Move d2l-filter demo component to Lit
