+++ demo/d2l-filter-dropdown/d2l-filter-dropdown.html
@@ -10,10 +10,10 @@
 			import '../../components/d2l-filter-dropdown/d2l-filter-dropdown.js';
 			import '../../components/d2l-filter-dropdown/d2l-filter-dropdown-category.js';
 			import '../../components/d2l-filter-dropdown/d2l-filter-dropdown-option.js';
-			import './d2l-demo-filter.js';
+			import './templated-filter-dropdown-demo.js';
 		</script>
 		<style>
-			d2l-demo-filter, d2l-filter-dropdown { display: inline-block; }
+			d2l-labs-templated-filter-dropdown-demo, d2l-filter-dropdown { display: inline-block; }
 		</style>
 	</head>
 	<body unresolved>
@@ -24,7 +24,7 @@
 
 			<d2l-demo-snippet>
 				<template>
-					<d2l-demo-filter></d2l-demo-filter>
+					<d2l-labs-templated-filter-dropdown-demo></d2l-labs-templated-filter-dropdown-demo>
 				</template>
 			</d2l-demo-snippet>
